refactor : [STEP2] 문자열 덧셈 계산기
